+++ src/language.cpp
@@ -115,7 +115,7 @@ void language_class::load(std::string file_name)
                         temp_char_UTF32 = data_line[count];
                         if(temp_char_UTF32 != ' ') temp_string_key += temp_char_UTF32;
                         count++;
-                        if(count > data_line.length()) (temp_char_UTF32 = ' ');
+                        if(count > (int)data_line.length()) (temp_char_UTF32 = ' ');
                     }
                     while((temp_char_UTF32 == ' ') || (temp_char_UTF32 == '='))
                     {
